backend/requirements.txt
fastapi
uvicorn
celery
pika
neo4j
musicbrainzngs
requests
pydantic
sqlitedict
python-multipart
python-dotenv
svgwrite